+++ src/main/java/io/xlorey/FluxLoader/client/core/Core.java
@@ -21,6 +21,6 @@ public class Core {
 
         PluginManager.loadPlugins(true);
 
-        UIManager.init();
+        WidgetManager.init();
     }
 }